Texture engine was not being initialized properly.
Texture() was reverted to no longer take (game,opt), it just takes (opt).  This caused initialization failure in voxpopuli because voxel-texture tries to get game.THREE and game.ao, but there was no game member on the opts object (in fact, the opts object was game itself and so it was completely broken).