Create container for game-shell, remove canvas z-index hack
The game-shell container is now created in createContainer(), ensuring
it is the first element in the document, then passed in the 'element'
shell option. If omitted, game-shell will create its own container, but
this will occur after plugins are loaded, which makes it difficult to
provide UI elements on top of the canvas since it is later in the DOM
tree, which was previously fixed by lowering the z-index but that causes
issues with fullscreen mode (see https://github.com/deathcap/voxel-fullscreen/issues/1).
Creating the container ourselves solves all these problems.